backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
alembic
psycopg2-binary
PyJWT[crypto]
passlib[bcrypt]
python-dotenv
openai
httpx
pydantic